refactor: extract AI match endpoints into wvImportAIController
Move 7 AI-match handlers (startAIMatch, getAIMatchStatus,
cancelAIMatchEndpoint, dbSearchOneRegion, geocodeMatch, resetMatch,
aiMatchOneRegion) plus their AI-service imports
(startAIMatching, getAIMatchProgress, cancelAIMatch,
aiMatchSingleRegion, dbSearchSingleRegion, geocodeMatchRegion,
isOpenAIAvailable) out of the worldViewImportController monolith
into a dedicated wvImportAIController file. The monolith
re-exports the 7 handlers.

Pure refactor — no behavior change, byte-identical function bodies.
Continues the spine domain-split (ADR-0009).
